mv disconnection codes/reasons to WsEndpoint
