+++ test/integration/duplicate-connections-are-closed.test.js
@@ -1,8 +1,7 @@
 const { waitForEvent } = require('streamr-test-utils')
 
-const { startEndpoint } = require('../../src/connection/WsEndpoint')
+const { startEndpoint, disconnectionReasons } = require('../../src/connection/WsEndpoint')
 const { PeerInfo } = require('../../src/connection/PeerInfo')
-const { disconnectionReasons } = require('../../src/messageTypes')
 const { LOCALHOST } = require('../util')
 
 describe('duplicate connections are closed', () => {
