+++ test/integration/killing-dead-connections.test.js
@@ -2,11 +2,10 @@
 const allSettled = require('promise.allsettled')
 const { waitForEvent } = require('streamr-test-utils')
 
-const { startEndpoint } = require('../../src/connection/WsEndpoint')
+const { startEndpoint, disconnectionReasons, disconnectionCodes } = require('../../src/connection/WsEndpoint')
 const { PeerInfo } = require('../../src/connection/PeerInfo')
 const { events } = require('../../src/connection/WsEndpoint')
 const { LOCALHOST } = require('../util')
-const { disconnectionReasons, disconnectionCodes } = require('../../src/messageTypes')
 
 describe('check and kill dead connections', () => {
     let node1
