+++ test/integration/tracker-node-reconnect-instructions.test.js
@@ -7,7 +7,7 @@ const TrackerServer = require('../../src/protocol/TrackerServer')
 const Node = require('../../src/logic/Node')
 const TrackerNode = require('../../src/protocol/TrackerNode')
 const endpointEvents = require('../../src/connection/WsEndpoint').events
-const { disconnectionReasons } = require('../../src/messageTypes')
+const { disconnectionReasons } = require('../../src/connection/WsEndpoint')
 
 /**
  * This test verifies that tracker can send instructions to node and node will connect and disconnect based on the instructions
