+++ test/integration/ws-endpoint.test.js
@@ -3,10 +3,9 @@ const { waitForEvent, wait } = require('streamr-test-utils')
 
 const { LOCALHOST } = require('../util')
 const endpointEvents = require('../../src/connection/WsEndpoint').events
-const { startEndpoint } = require('../../src/connection/WsEndpoint')
+const { startEndpoint, disconnectionCodes } = require('../../src/connection/WsEndpoint')
 const { PeerInfo } = require('../../src/connection/PeerInfo')
 const { startTracker } = require('../../src/composition')
-const { disconnectionCodes } = require('../../src/messageTypes')
 
 describe('ws-endpoint', () => {
     const MAX = 5
